+++ proyecto2-raytracer/src/geometry/mod.rs
@@ -1,0 +1,7 @@
+//! Geometry primitives used for tracing (solid cubes, textured variants, etc.).
+
+pub mod solid_block;
+pub mod textured_block;
+
+pub use solid_block::SolidBlock;
+pub use textured_block::TexturedBlock;
