+++ proyecto2-raytracer/src/math/mod.rs
@@ -1,0 +1,5 @@
+//! Math utilities (vectors, rays, etc.)
+
+pub mod vec3;
+
+pub use vec3::Vec3;
