+++ proyecto2-raytracer/src/rendering/mod.rs
@@ -1,0 +1,7 @@
+//! Rendering pipeline and lighting helpers.
+
+pub mod lighting;
+pub mod pipeline;
+
+pub use lighting::{Skybox, Tex};
+pub use pipeline::{Assets, WorldKind, render};
