+++ proyecto2-raytracer/src/scene/mod.rs
@@ -1,0 +1,5 @@
+//! Scene-level data structures (materials, traits for hittable objects).
+
+pub mod material;
+
+pub use material::{Intersectable, MaterialParams};
